Modern Mainframe Development - Tom Taulli Even as spending on digital transformation continues to skyrocket, mainframes nevertheless have major advantages for global enterprises. These systems still process huge amounts of information and allow for highly secure transactions. In this practical book, author Tom Taulli shows software developers how to pursue a hybrid approach by integrating traditional mainframes and applications with modern digital systems.
By the end of this book, you'll have a solid understanding of the mainframe architecture and ecosystem, including core concepts and technologies such as COBOL, REXX, JCL, Db2, VSAM, and CICS. You'll also learn how to blend in newer technologies such as the cloud, AI and machine learning, and microservices. This handbook is indispensable for enterprises looking to thrive in the new digital world.

+ Learn strategies and approaches for mainframe DevOps
+ Create, update, and maintain existing mainframe code, and analyze and resolve common errors
+ Apply modern approaches to the mainframe, including microservices, APIs, cloud, and AI and machine learning
+ Work with datasets and databases and put together effective reports
+ Understand how to work with modern cloud systems, like AWS, for pursuing data migration