Learning TypeScript - Josh Goldberg TypeScript has conquered the world of JavaScript: it's one of the world's fastest growing and most popular languages across developer surveys, widely used in consumer and business companies alike, and frequently credited for helping massive web applications scale. But what is TypeScript? How does it work, why does it work, and how can we use it?
Learning TypeScript takes beginner to intermediate JavaScript programmers from knowing nothing about types or a type system to full mastery of the fundamentals of TypeScript. It's more than a means to find bugs and typos--it's a useful system for declaring the way our JavaScript should work and helping us stick to it.

You'll learn how TypeScript:
+ interacts with JavaScript
+ analyzes and understands code
+ augments your existing development pattern
+ helps you document your code
+ works with IDEs to provide refactoring tools
+ assists local development in refactoring code
+ helps you develop more quickly with fewer bugs