Reinforcement Learning for Finance - Yves Hilpisch Reinforcement learning (RL) has led to several breakthroughs in AI. The use of the Q-learning (DQL) algorithm alone has helped people develop agents that play arcade games and board games at a superhuman level. More recently, RL, DQL, and similar methods have gained popularity in publications related to financial research.
This book is among the first to explore the use of reinforcement learning methods in finance.

Author Yves Hilpisch, founder and CEO of The Python Quants, provides the background you need in concise fashion. ML practitioners, financial traders, portfolio managers, strategists, and analysts will focus on the implementation of these algorithms in the form of self-contained Python code and the application to important financial problems.

This book covers:
+ Reinforcement learning
+ Deep Q-learning
+ Python implementations of these algorithms
+ How to apply the algorithms to financial problems such as algorithmic trading, dynamic hedging, and dynamic asset allocation

This book is the ideal reference on this topic. You'll read it once, change the examples according to your needs or ideas, and refer to it whenever you work with RL for finance.

Dr. Yves Hilpisch is founder and CEO of The Python Quants, a group that focuses on the use of open source technologies for financial data science, AI, asset management, algorithmic trading, and computational finance.